+++ quarkus-cafe-core/src/main/java/com/redhat/quarkus/cafe/infrastructure/OrderRepository.java
@@ -8,9 +8,4 @@ import javax.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
 public class OrderRepository implements PanacheMongoRepository<Order> {
-
-    @Override
-    public void persist(Order order) {
-        order.id = ObjectId.get();
-    }
 }
